Tests - Junit - told to do  after completion
Tests carried out after completion as keith stated this was to be
done...was not originally in the instructions, and keith has no t
specifically told us what to test. Thus i have therefore only tested the
code i created.

all tested and works - translatorTest has string Line change (in
translator class) just for that test.

Translator class getInstruction() has been modified following the tests
to make it flow better and not repeat an if check